requirements.txt
Django==3.2.5
django-import-export==2.5.0
django-simple-history==3.0.0
django-treebeard==4.5.1
django-admin-rangefilter==0.8.1
django-cors-headers==3.7.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.27
dj-rest-auth==2.1.8
djangorestframework==3.12.4
django-allauth==0.44.0
drf-yasg==1.20.0
psycopg2==2.9.1
django-extensions==3.1.3
drf-extensions==0.7.0
# dev
isort==5.9.2
ipython==7.25.0
pre-commit==2.13.0
Pillow==8.3.1
